Version 1.0 -In Admin Section added button to change status of seen messages

+++ src/main/resources/schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS `contact_msg` (
-                                             `contact_id` int AUTO_INCREMENT PRIMARY KEY,
+                                             `contact_id` long AUTO_INCREMENT PRIMARY KEY,
                                              `name` varchar(100) NOT NULL,
                                              `mobile_num` varchar(10) NOT NULL,
                                              `email` varchar(100) NOT NULL,
